New API-url added to access 7 days forecast
Inside the function wordTemperatue I called the function getWeeklyForecast which will fetch the coordinates (lat and lon) when we get a response from the API. Then I go to the  getWeeklyForecast and create the API-url of the new API, we are sending the coordinates (lat, lon) + aplyKey and the units of metric, then I ask axios to fetch me that apiUlr, and once is I get the response go and display the forescast. The result is an array with 8 elements and each one will have temp.